+++ src/main/java/com/fourweekdays/fourweekdays/announcement/controller/AnnouncementController.java
@@ -9,9 +9,7 @@ import com.fourweekdays.fourweekdays.member.model.UserAuth;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
