+++ src/main/java/com/fourweekdays/fourweekdays/announcement/model/entity/Announcement.java
@@ -3,8 +3,6 @@ package com.fourweekdays.fourweekdays.announcement.model.entity;
 import com.fourweekdays.fourweekdays.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
-
 
 
 @Entity
