+++ src/main/java/com/fourweekdays/fourweekdays/franchise/model/dto/request/FranchiseCreateDto.java
@@ -3,7 +3,6 @@ package com.fourweekdays.fourweekdays.franchise.model.dto.request;
 import com.fourweekdays.fourweekdays.common.vo.Address;
 import com.fourweekdays.fourweekdays.franchise.model.entity.FranchiseStatus;
 import com.fourweekdays.fourweekdays.franchise.model.entity.FranchiseStore;
-import jakarta.persistence.*;
 import lombok.*;
 
 @Getter
