+++ src/main/java/com/fourweekdays/fourweekdays/image/service/UploadService.java
@@ -3,7 +3,6 @@ package com.fourweekdays.fourweekdays.image.service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.sql.SQLException;
 
 public interface UploadService {
     String upload(MultipartFile file) throws IOException;
