+++ src/main/java/com/fourweekdays/fourweekdays/member/controller/MemberController.java
@@ -1,23 +1,16 @@
 package com.fourweekdays.fourweekdays.member.controller;
 
-import com.fourweekdays.fourweekdays.announcement.model.dto.response.AnnouncementReadDto;
 import com.fourweekdays.fourweekdays.common.BaseResponse;
-import com.fourweekdays.fourweekdays.inbound.model.dto.response.InboundReadDto;
 import com.fourweekdays.fourweekdays.member.model.dto.*;
-import com.fourweekdays.fourweekdays.member.model.entity.AuthStatus;
-import com.fourweekdays.fourweekdays.member.model.entity.MemberRole;
 import com.fourweekdays.fourweekdays.member.service.MemberService;
-import jakarta.validation.Valid;
-import com.fourweekdays.fourweekdays.product.dto.request.ProductUpdateDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @Tag(name = "직원 관리")
 @RestController
@@ -74,7 +67,7 @@ public class MemberController {
 
     //이메일 중복체크 기능
     @PostMapping("/check-email")
-    public  ResponseEntity<BaseResponse<String>> checkEmail(@RequestBody MemberEmailCheckDto dto) {
+    public ResponseEntity<BaseResponse<String>> checkEmail(@RequestBody MemberEmailCheckDto dto) {
         memberService.checkEmailDuplicate(dto.getEmail());
         return ResponseEntity.ok(BaseResponse.success("사용 가능한 이메일입니다."));
     }
