+++ src/main/java/com/fourweekdays/fourweekdays/member/model/dto/MemberLoginResponseDto.java
@@ -2,7 +2,6 @@ package com.fourweekdays.fourweekdays.member.model.dto;
 
 import com.fourweekdays.fourweekdays.member.model.UserAuth;
 import com.fourweekdays.fourweekdays.member.model.entity.MemberRole;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
