+++ src/main/java/com/fourweekdays/fourweekdays/member/repository/MemberRepositoryCustom.java
@@ -7,8 +7,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.List;
 
 public interface MemberRepositoryCustom {
