+++ src/main/java/com/fourweekdays/fourweekdays/member/repository/MemberRepositoryCustomImpl.java
@@ -13,7 +13,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 import static com.fourweekdays.fourweekdays.member.model.entity.QMember.member;
