+++ src/main/java/com/fourweekdays/fourweekdays/order/filter/FranchiseApiKeyFilter.java
@@ -3,9 +3,6 @@ package com.fourweekdays.fourweekdays.order.filter;
 import com.fourweekdays.fourweekdays.franchise.model.entity.FranchiseStatus;
 import com.fourweekdays.fourweekdays.franchise.model.entity.FranchiseStore;
 import com.fourweekdays.fourweekdays.franchise.repository.FranchiseRepository;
-import com.fourweekdays.fourweekdays.vendor.model.entity.Vendor;
-import com.fourweekdays.fourweekdays.vendor.model.entity.VendorStatus;
-import com.fourweekdays.fourweekdays.vendor.repository.VendorRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
