+++ src/main/java/com/fourweekdays/fourweekdays/order/model/dto/request/OrderProductDto.java
@@ -2,7 +2,6 @@ package com.fourweekdays.fourweekdays.order.model.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.Getter;
 
 @Getter
