+++ src/main/java/com/fourweekdays/fourweekdays/order/resolver/FranchiseArgumentResolver.java
@@ -1,6 +1,5 @@
 package com.fourweekdays.fourweekdays.order.resolver;
 
-import com.fourweekdays.fourweekdays.asn.annotation.AuthenticatedVendor;
 import com.fourweekdays.fourweekdays.order.anootation.AuthenticatedFranchise;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
