+++ src/main/java/com/fourweekdays/fourweekdays/product/controller/ProductController.java
@@ -13,8 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @Tag(name = "상품 기능")
 @RestController
 @RequiredArgsConstructor
