+++ src/main/java/com/fourweekdays/fourweekdays/product/model/dto/request/ProductCreateDto.java
@@ -2,9 +2,6 @@ package com.fourweekdays.fourweekdays.product.model.dto.request;
 
 import com.fourweekdays.fourweekdays.product.model.entity.Product;
 import com.fourweekdays.fourweekdays.product.model.entity.ProductStatus;
-import com.fourweekdays.fourweekdays.product.model.Product;
-import com.fourweekdays.fourweekdays.product.model.ProductStatus;
-import com.fourweekdays.fourweekdays.vendor.model.entity.Vendor;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
