+++ src/main/java/com/fourweekdays/fourweekdays/purchaseorder/controller/PurchaseOrderController.java
@@ -13,8 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @Tag(name = "발주 기능")
 @RequestMapping("/api/purchase-orders")
 @RequiredArgsConstructor
