+++ src/main/java/com/fourweekdays/fourweekdays/warehouse/exception/WarehouseExceptionType.java
@@ -3,7 +3,6 @@ package com.fourweekdays.fourweekdays.warehouse.exception;
 import com.fourweekdays.fourweekdays.global.exception.ExceptionType;
 import org.springframework.http.HttpStatus;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 public enum WarehouseExceptionType implements ExceptionType {
