+++ src/test/java/com/fourweekdays/fourweekdays/FourweekdaysApplicationTests.java
@@ -1,7 +1,6 @@
 package com.fourweekdays.fourweekdays;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
 //@SpringBootTest
 class FourweekdaysApplicationTests {
