+++ src/test/java/com/fourweekdays/fourweekdays/purchaseorder/service/PurchaseOrderServiceTest.java
@@ -2,9 +2,6 @@ package com.fourweekdays.fourweekdays.purchaseorder.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 //@SpringBootTest
 class PurchaseOrderServiceTest {
